NHIRS-18 - Initial Earthquake template with example (#39)
* NHIRS-18: Add EQ template and example

* NHIRS-18: Split templates into new module separated by hazard type

* NHIRS-18: Enable hazard_raster attribute_label to be configured
