+++ examples/earthquake/v1/gascoyne/input/continental_crust/mmi_mean.hdr
@@ -1,0 +1,14 @@
+BYTEORDER  LSBFIRST
+LAYOUT  BIL
+NROWS  217
+NCOLS  235
+NBANDS  1
+NBITS  32
+BANDROWBYTES  940.0
+TOTALROWBYTES  940.0
+PIXELTYPE  FLOAT
+ULXMAP  113.53333333333333
+ULYMAP  -21.95
+XDIM  0.008333333333333333
+YDIM  0.008333333333333333
+NODATA  999.0
