+++ examples/earthquake/v1/gascoyne/input/exposure.csv
