+++ examples/earthquake/v1/gascoyne/output/.gitignore
@@ -1,0 +1,2 @@
+*
+*/
