+++ hazimp/templates/constants.py
@@ -1,0 +1,27 @@
+DEFAULT = 'default'
+
+WINDV3 = 'wind_v3'
+WINDV4 = 'wind_v4'
+WINDV5 = 'wind_v5'
+WINDNC = 'wind_nc'
+
+EARTHQUAKEV1 = 'earthquake_v1'
+
+FLOODFABRICV2 = 'flood_fabric_v2'
+FLOODCONTENTSV2 = 'flood_contents_v2'
+
+REP_VAL_NAME = 'replacement_value_label'
+
+TEMPLATE = 'template'
+SAVE = 'save'
+SAVEAGG = 'save_agg'
+LOADWINDTCRM = 'load_wind'
+HAZARDRASTER = 'hazard_raster'
+CALCSTRUCTLOSS = 'calc_struct_loss'
+CALCCONTLOSS = 'calc_cont_loss'
+PERMUTATION = 'exposure_permutation'
+VULNFILE = 'vulnerability_filename'
+VULNSET = 'vulnerability_set'
+AGGREGATION = 'aggregation'
+AGGREGATE = 'aggregate'
+TABULATE = 'tabulate'
